feat: refactorización de gestión de géneros y selector reutilizable
- Refactoricé la lógica de creación, actualización y eliminación de géneros en un solo componente unificado.
- Implementé el componente reutilizable <app-list-genres> que emite el id del género seleccionado vía Output.
- Añadí el método onGenreChange para emitir correctamente el id seleccionado.
- Agregué la carpeta test_genre para pruebas manuales y validación del Output del selector de géneros.
- Mejoras de UI/UX: placeholder Genre visible y estilizado correctamente, tamaño y estilos del select unificados para uso consistente.
