+++ src/app/features/genre/genre.routes.ts
@@ -1,8 +1,9 @@
+import { TestListGenreComponent } from "./pages/test_genre/test-list-genre.component";
 import { Routes } from "@angular/router";
 
 export const GENRE_ROUTES: Routes = [
   {
     path: '',
-    loadComponent: () => import('./pages/list_genres/list-genres.component').then(m => m.ListGenresComponent)
+    loadComponent: () => import('./pages/test_genre/test-list-genre.component').then(m => m.TestListGenreComponent)
   }
 ];
