+++ src/app/features/genre/pages/crud_genre/genre.service.spec.ts
@@ -1,4 +1,4 @@
-import { GenreService } from './../../../../services/genre.service';
+import { GenreService } from '../../../../services/genre.service';
 import { TestBed } from '@angular/core/testing';
 
 describe('GenreService', () => {
