+++ src/app/features/genre/pages/list_genres/list-genres.component.css
@@ -1,6 +1,10 @@
 .genre-select {
+  background-image: url("data:image/svg+xml;utf8,<svg fill='black' height='16' viewBox='0 0 24 24' width='16' xmlns='http://www.w3.org/2000/svg'><path d='M7 10l5 5 5-5z'/></svg>");
+  background-repeat: no-repeat;
+  background-position: right 12px center;
+  background-size: 32px 32px;
   width: 394px;
-  height: 27px;
+  height: 37px;
   background-color: #F3F3F3;
   border-radius: 5px;
   border: none;
@@ -9,10 +13,10 @@
   padding: 8px 12px;
   outline: none;
   appearance: none;
-  box-sizing: border-box; /* importante para que el padding no rompa el tamaño */
+  box-sizing: border-box; 
 }
 
-/* Si quieres mantener los estilos para los options */
+
 .genre-select option {
   font-family: 'Manjari', sans-serif;
   font-size: 16px;
@@ -20,7 +24,5 @@
   color: #000;
 }
 
-/* Placeholder gris para opción por defecto */
-.genre-select:invalid {
-  color: rgba(114,111,111,0.28); 
-}
+
+
