+++ src/app/features/genre/pages/list_genres/list-genres.component.html
@@ -1,21 +1,13 @@
 <form>
   <select
     class="genre-select"
-    [ngStyle]="{
-      'background-color': '#F3F3F3',
-      'border-radius': '5px',
-      'font-family': 'Manjari, sans-serif',
-      'font-size': '15px',
-      'border': 'none',
-      'padding': '8px 12px',
-      'color': selectedGenre ? '#000' : 'rgba(114,111,111,0.28)'
-    }"
     [(ngModel)]="selectedGenre"
     name="genre"
-    style="outline: none; appearance: none;"
+    (change)="onGenreChange()"
+    required
   >
     <option value="" disabled selected hidden>Género</option>
-    <option *ngFor="let genre of genres" [value]="genre.name">
+    <option *ngFor="let genre of genres" [value]="genre.id">
       {{ genre.name }}
     </option>
   </select>
